Newsapp/requirements.txt
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
Automat==20.2.0
autopep8==1.5
beautifulsoup4==4.8.2
bleach==3.1.1
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
constantly==15.1.0
crontab==0.22.6
cryptography==2.8
cssselect==1.1.0
decorator==4.4.2
dicttoxml==1.7.4
Django==3.0.3
django-crontab==0.7.1
django-ipware==2.1.0
django-qr-code==1.1.0
djangorestframework==3.11.0
docutils==0.16
future==0.18.2
geocoder==1.38.1
geoip2==3.0.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
ip2geotools==0.1.5
IP2Location==8.4.1
isort==4.3.21
keyring==21.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
maxminddb==1.5.2
mccabe==0.6.1
packaging==20.3
parsel==1.5.2
Pillow==7.0.0
pip-review==1.0
pkginfo==1.5.0.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyquery==1.4.1
pytz==2019.3
pywin32-ctypes==0.2.0
qrcode==6.1
queuelib==1.5.0
ratelim==0.1.6
readme-renderer==25.0
requests==2.23.0
requests-toolbelt==0.9.1
Scrapy==2.0.1
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
tqdm==4.43.0
twine==3.1.1
Twisted==20.3.0
typed-ast==1.4.1
typing==3.7.4.1
urllib3==1.25.8
w3lib==1.21.0
webencodings==0.5.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.0.2
